Fix UI layout issues
Created a "change Players" button and placed it with the reset button
under the gameboard. Styled both in the same style to be visible against
the gradient background. Added margins to the stats div elements, and
updated JS code to apply appropriate styling when the reset and submit
buttons are pressed.